added a grid component for new collections, popular collections, and collections to make the code cleaner
