+++ src/pages/CollectionsPage.jsx
@@ -1,12 +1,10 @@
 import React, { useEffect, useState } from "react";
-import { Link } from "react-router-dom";
-import Skeleton from "../components/ui/Skeleton";
 import axios from "axios";
+import CollectionGrid from "../components/collection/CollectionGrid";
 
 export default function CollectionsPage() {
   const [collections, setCollections] = useState([]);
   const [loading, setLoading] = useState(true);
-  const [visible, setVisible] = useState(12)
 
   useEffect(() => {
     async function fetchData() {
@@ -32,67 +30,7 @@ export default function CollectionsPage() {
     <div className="container">
       <div className="row">
         <h1 className="collections-page__title">Collections</h1>
-        <div className="collections__body">
-          {loading
-            ? new Array(6).fill(0).map((_, i) => (
-                <div key={i} className="collection-column">
-                  <div className="collection">
-                    <Skeleton width="100%" height="200px" borderRadius="12px" />
-                    <div className="collection__info">
-                      <Skeleton width="60%" height="1rem" borderRadius="4px" />
-                      <div className="collection__stats">
-                        <Skeleton
-                          width="80px"
-                          height="1rem"
-                          borderRadius="4px"
-                        />
-                        <Skeleton
-                          width="80px"
-                          height="1rem"
-                          borderRadius="4px"
-                        />
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              ))
-            : collections.slice(0, visible).map((item) => (
-                <div key={item.collectionId} className="collection-column">
-                  <Link to="/collection/${item.collectionId}" className="collection">
-                    <img
-                      src={item.imageLink}
-                      alt=""
-                      className="collection__img"
-                    />
-                    <div className="collection__info">
-                      <h3 className="collection__name">{item.title}</h3>
-                      <div className="collection__stats">
-                        <div className="collection__stat">
-                          <span className="collection__stat__label">Floor</span>
-                          <span className="collection__stat__data">
-                            {parseFloat(item.floor).toFixed(2)} ETH
-                          </span>
-                        </div>
-                        <div className="collection__stat">
-                          <span className="collection__stat__label">
-                            Total Volume
-                          </span>
-                          <span className="collection__stat__data">
-                            {item.totalVolume} ETH
-                          </span>
-                        </div>
-                      </div>
-                    </div>
-                  </Link>
-                </div>
-              ))}
-        </div>
-        {!loading && visible < collections.length && (
-        <button 
-        className="collections-page__button"
-        onClick={() => setVisible((prev) => prev + 6)}
-        >Load more</button>
-        )}
+        <CollectionGrid collections={collections} loading={loading} loadMore />
       </div>
     </div>
   );
